requirements.txt
gradio[mcp]>=5.32.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.0.0
supabase>=2.0.0
python-multipart>=0.0.6
pydantic>=2.0.0
numpy>=1.24.0
aiohttp>=3.12.0 